request: checkTimeout() after identified
We need to do another checkTimeout() after the request
is identified to double check that we did not consume our
request budget in the connection init handshake.

We also need to update the `elapsed` time so that when we
make our outbound request we set a proper `ttl-elapsed` timeout
on the per attempt request.